# specs/001-multilingual-kiosk-quiz/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Multilingual Touchscreen Quiz API
  version: 0.1.0
  description: API contract for kiosk quiz flow, leaderboard, and public exports.
servers:
  - url: https://quiz.example.org
paths:
...
  /api/sessions/{sessionId}/result:
    get:
      summary: Get final result for session
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Final result summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResultResponse"
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    SessionResultResponse:
      type: object
      properties:
        score:
          type: integer
        maxScore:
          type: integer
        percentage:
          type: number
          format: float
        leaderboard:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
      required: [score, maxScore, percentage, leaderboard]
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        nickname:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        percentage:
          type: number
          format: float
      required: [rank, nickname, score, maxScore, percentage]